Remove requirement for surf login to have run before unit tests
